// Zybach.Web/src/app/shared/generated/model/support-ticket-simple-dto.ts
/**
 * Zybach
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { SupportTicketPrioritySimpleDto } from '././support-ticket-priority-simple-dto';
import { WellMinimalDto } from '././well-minimal-dto';
import { SupportTicketStatusSimpleDto } from '././support-ticket-status-simple-dto';
import { SensorSimpleDto } from '././sensor-simple-dto';
import { UserSimpleDto } from '././user-simple-dto';

export class SupportTicketSimpleDto { 
    SupportTicketID?: number;
    DateCreated?: string;
    DateUpdated?: string;
    CreatorUserID?: number;
    AssigneeUserID?: number;
    WellID?: number;
    SensorID?: number;
    SupportTicketStatusID?: number;
    SupportTicketPriorityID?: number;
    SupportTicketTitle?: string;
    SupportTicketDescription?: string;
    Well?: WellMinimalDto;
    Sensor?: SensorSimpleDto;
    CreatorUser?: UserSimpleDto;
    AssigneeUser?: UserSimpleDto;
    Status?: SupportTicketStatusSimpleDto;
    Priority?: SupportTicketPrioritySimpleDto;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}